Add bearer token validation to penalty endpoints

+++ src/main/kotlin/de/vexxes/routes/penalty/DeletePenalty.kt
@@ -1,5 +1,6 @@
 package de.vexxes.routes.penalty
 
+import de.vexxes.authorization.ValidateBearerToken
 import de.vexxes.domain.model.ApiResponse
 import de.vexxes.domain.model.Endpoint
 import de.vexxes.domain.repository.Repository
@@ -10,29 +11,36 @@ import io.ktor.server.routing.*
 
 fun Route.deletePenalty(
     app: Application,
-    repository: Repository
+    repository: Repository,
+    validateBearerToken: ValidateBearerToken
 ) {
-    put(Endpoint.DeletePenalty.path) {
-        try {
-            val penaltyId = call.parameters["penaltyId"]
-            val response = repository.deletePenalty(penaltyId = penaltyId)
 
-            if (response) {
-                call.respond(
-                    message = ApiResponse(
-                        success = true,
-                        message = "Successfully deleted!"
-                    ),
-                    status = HttpStatusCode.OK
-                )
-            } else {
-                call.respond(
-                    message = ApiResponse(success = false),
-                    status = HttpStatusCode.BadRequest
-                )
+    put(Endpoint.DeletePenalty.path) {
+        if (validateBearerToken.validateAdmin(call.request.headers["Authorization"].toString())) {
+            try {
+                val penaltyId = call.parameters["penaltyId"]
+                val response = repository.deletePenalty(penaltyId = penaltyId)
+
+                if (response) {
+                    call.respond(
+                        message = ApiResponse(
+                            success = true,
+                            message = "Successfully deleted!"
+                        ),
+                        status = HttpStatusCode.OK
+                    )
+                } else {
+                    call.respond(
+                        message = ApiResponse(success = false),
+                        status = HttpStatusCode.BadRequest
+                    )
+                }
+            } catch (e: Exception) {
+                app.log.info("DELETE PENALTY INFO ERROR: $e")
             }
-        } catch (e: Exception) {
-            app.log.info("DELETE PENALTY INFO ERROR: $e")
+        } else {
+            app.log.info("authentication failed")
+            call.respond(HttpStatusCode.Unauthorized)
         }
     }
 }
